// src/app/shared/models/user.ts
export interface User {
  uid?: string;
  email?: string;
  displayName: string;
  photoURL: string;
  emailVerified: boolean;
  adventista?: boolean;
  camiseta?: string;
  patente?: string;
  igreja?: string;
  dataNascimento?: string;
  pontos?: number;
  sangue?: string;
  cadastrado?: string;
}